Add error objects to represent timeouts
Re-organize exception hierarchy in a backwards-compatible manner.

Introduce the new base exception class called `BaseError`. It will be a
new parent for both `SetupError` and `Error`. This doesn't change
anything for clients of the code who were previously recsuing from the
exceptions separately.

We'll also introduce separate branch of classes for timeouts:
`FirebaseCloudMessenger::Timeout` will be a parent to
`FirebaseCloudMessenger::ConnectTimeout` and
`FirebaseCloudMessenger::ReadTimeout` that will represent the connect
and read timeout errors, respectively.